requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0
markdown==3.5.1
Pygments==2.17.2
bleach==6.1.0
waitress==3.0.0
gunicorn
python-slugify==8.0.1
celery[redis]==5.3.4
redis==4.6.0
user-agents==2.2.0
colorama==0.4.6
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
